+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-06-01 21:37:16
-// Commit: fix(ui): prevent double tap duplicate action in ui
+// Updated at: 2026-06-01 22:56:27
+// Commit: feat(ui): add bi-directional event bus handler for ui
 
 class uiComponentState {
-  final String commitRef = '1677';
-  final String lastMessage = 'fix(ui): prevent double tap duplicate action in ui';
-  final DateTime timestamp = DateTime.parse('2026-06-01T21:37:16+05:00');
+  final String commitRef = '1680';
+  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
+  final DateTime timestamp = DateTime.parse('2026-06-01T22:56:27+05:00');
 }